windowsuse/pyproject.toml
[project]
name = "windowsuse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "windows-use>=0.1.4",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.1",
]
